refactor: pluralization of `config` folder
+ some refactoring and sanitization
