+++ .env.example
@@ -4,7 +4,7 @@ TEST_GUILD_ID="TEST_GUILD_ID"
 BOT_OWNER_ID="YOUR_DISCORD_ID"
 
 # database
-DATABASE_HOST="database" # name of the docker-compose service hosting the database
+DATABASE_HOST="database" # if you use docker-compose, it should be the name of the service hosting the database
 DATABASE_PORT=3306
 DATABASE_NAME="tscord_bot"
 DATABASE_USER="tscord"
