+++ src/api/controllers/database.ts
@@ -4,7 +4,7 @@ import { Required } from "@tsed/schema"
 import { injectable } from "tsyringe"
 
 import { Authenticated } from "@api/middlewares"
-import { databaseConfig } from "@config"
+import { databaseConfig } from "@configs"
 import { Database } from "@services"
 import { BaseController } from "@utils/classes"
 import { formatDate, resolveDependencies } from "@utils/functions"
