+++ src/commands/General/invite.ts
@@ -2,7 +2,7 @@ import { Category } from "@discordx/utilities"
 import { CommandInteraction, EmbedBuilder } from "discord.js"
 import { Client } from "discordx"
 
-import { generalConfig } from "@config"
+import { generalConfig } from "@configs"
 import { Discord, Slash } from "@decorators"
 import { Guard } from "@guards"
 import { getColor } from "@utils/functions"
