+++ src/configs/general.ts
