+++ src/configs/index.ts
