+++ src/utils/functions/eval.ts
@@ -1,6 +1,6 @@
 import { Message } from "discord.js"
 
-import { generalConfig } from "@config"
+import { generalConfig } from "@configs"
 
 const clean = (text: any) => {
     if (typeof (text) === 'string') return text.replace(/`/g, '`' + String.fromCharCode(8203)).replace(/@/g, '@' + String.fromCharCode(8203))
