+++ src/utils/functions/localization.ts
@@ -1,4 +1,4 @@
-import { generalConfig } from "@config"
+import { generalConfig } from "@configs"
 import { L, loadedLocales, Locales, locales } from "@i18n"
 
 export const getLocalizedInfo = (target: 'NAME' | 'DESCRIPTION', localizationSource: TranslationsNestedPaths) => {
