+++ tsconfig.json
@@ -40,8 +40,8 @@
 			"@i18n": ["src/i18n"],
 			"@i18n/*": ["src/plugins/*/i18n"],
 
-			"@config": ["src/config"],
-			"@config/*": ["src/plugins/*/config"],
+			"@configs": ["src/configs"],
+			"@configs/*": ["src/plugins/*/configs"],
 
 			"@utils/classes": ["src/utils/classes"],
 			"@utils/classes/*": ["src/plugins/*/utils/classes"],
